// packages/packages/frontend/core/src/components/page-list/types.ts
export type TagMeta = {
  id: string;
  name: string;
  color: string;
  pageCount?: number;
  createDate?: Date | number;
  updatedDate?: Date | number;
  parentId?: string;
  depth?: number;
  hasChildren?: boolean;
  collapsed?: boolean;
  onToggleCollapse?: () => void;
};